gradeOps-new/ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
Pillow==10.1.0
